refactor(master): Remove C-Style memory allocation
 - The read now uses an InputPacket type to handle reads from clients.
 - Remove old C-Style memory allocation for inputpacket messages.
 - Simplify read logic to handle state change between HEADER and DATA.
