+++ src/chunkserver/master_connection.cc
@@ -38,6 +38,7 @@
 #include "chunkserver/bgjobs.h"
 #include "chunkserver/hddspacemgr.h"
 #include "chunkserver/network_main_thread.h"
+#include "common/input_packet.h"
 #include "common/loop_watchdog.h"
 #include "common/network_address.h"
 #include "common/output_packet.h"
@@ -46,7 +47,6 @@
 #include "config/cfg.h"
 #include "protocol/SFSCommunication.h"
 #include "protocol/cstoma.h"
-#include "protocol/input_packet.h"
 #include "protocol/matocs.h"
 #include "protocol/packet.h"
 #include "slogger/slogger.h"
