+++ src/common/input_packet.h
