+++ src/master/matocsserv.cc
@@ -41,6 +41,7 @@
 #include "common/datapack.h"
 #include "common/event_loop.h"
 #include "common/goal.h"
+#include "common/input_packet.h"
 #include "common/loop_watchdog.h"
 #include "common/massert.h"
 #include "common/output_packet.h"
@@ -59,7 +60,6 @@
 #include "master/personality.h"
 #include "protocol/SFSCommunication.h"
 #include "protocol/cstoma.h"
-#include "protocol/input_packet.h"
 #include "protocol/matocs.h"
 #include "protocol/packet.h"
 #include "slogger/slogger.h"
